Moved the shared filter engine into admin-x-framework
- the filter AST/codec system and the value-source engine were shared across the posts app's comments and members domains, which dissolve into apps/admin in separate PRs; a workspace package can't import from apps/admin, so the shared code had to move to a neutral home both can reach
- it's coupled to react-query (the value sources wrap the framework api hooks) and to the router, so framework is the correct home, not shade
- exposed as @tryghost/admin-x-framework/filters and /filter-sources subpaths, mirroring the virtual-list primitive; get-site-timezone moved too, since comments/members/post-analytics all use it
- the per-domain field definitions stay with their views and move later with their domain — only the shared primitives and value-source engine were extracted
