+++ apps/admin-x-framework/src/filter-sources/use-label-value-source.ts
@@ -1,10 +1,10 @@
-import {Label, LabelsResponseType, useBrowseLabelsInfinite} from '@tryghost/admin-x-framework/api/labels';
+import {Label, LabelsResponseType, useBrowseLabelsInfinite} from '../api/labels';
 import {RemoteProbeState, ValueSourceHookOptions} from './create-remote-value-source';
 import {ValueSource, ValueSourceParams, ValueSourceState} from '@tryghost/shade/patterns';
 import {buildQuotedListFilter, filterOptionsByQuery, mergeFilterOptions} from './utils';
 import {createGhostBrowseValueSource} from './create-ghost-browse-value-source';
 import {createHybridValueSource} from './create-hybrid-value-source';
-import {escapeNqlString} from '@src/views/filters/filter-normalization';
+import {escapeNqlString} from '../filters/filter-normalization';
 
 const LABEL_PAGE_LIMIT = '100';
