+++ apps/admin-x-framework/src/filter-sources/use-member-value-source.ts
@@ -1,4 +1,4 @@
-import {Member, MembersInfiniteResponseType, useBrowseMembersInfinite} from '@tryghost/admin-x-framework/api/members';
+import {Member, MembersInfiniteResponseType, useBrowseMembersInfinite} from '../api/members';
 import {ValueSource} from '@tryghost/shade/patterns';
 import {createGhostBrowseValueSource} from './create-ghost-browse-value-source';
