+++ apps/admin-x-framework/src/filter-sources/use-post-resource-value-source.ts
@@ -1,9 +1,9 @@
-import {Page, PagesResponseType, useBrowsePagesInfinite} from '@tryghost/admin-x-framework/api/pages';
-import {Post, PostsResponseType, useBrowsePostsInfinite} from '@tryghost/admin-x-framework/api/posts';
+import {Page, PagesResponseType, useBrowsePagesInfinite} from '../api/pages';
+import {Post, PostsResponseType, useBrowsePostsInfinite} from '../api/posts';
 import {ValueSource} from '@tryghost/shade/patterns';
 import {createCombinedValueSource} from './create-combined-value-source';
 import {createGhostBrowseValueSource} from './create-ghost-browse-value-source';
-import {escapeNqlString} from '@src/views/filters/filter-normalization';
+import {escapeNqlString} from '../filters/filter-normalization';
 
 function buildPublishedFilter(query: string) {
     return query ? `status:published+title:~${escapeNqlString(query)}` : 'status:published';
