+++ apps/admin-x-framework/src/nql-lang.d.ts
@@ -1,0 +1,11 @@
+declare module '@tryghost/nql-lang' {
+    interface ParseOptions {
+        preserveRelativeDates?: boolean;
+    }
+
+    const nql: {
+        parse(input: string, options?: ParseOptions): unknown;
+    };
+
+    export default nql;
+}
