+++ apps/admin-x-framework/src/utils/get-site-timezone.ts
@@ -1,4 +1,4 @@
-import type {Setting} from '@tryghost/admin-x-framework/api/settings';
+import type {Setting} from '../api/settings';
 
 export const getSiteTimezone = (settings: ReadonlyArray<Setting>): string => {
     for (const setting of settings) {
