+++ apps/admin-x-framework/test/setup.ts
@@ -1,4 +1,4 @@
 /// <reference types="vitest/globals" />
 import '@testing-library/jest-dom';
 
-// This file ensures TypeScript knows about vitest globals 
+// This file ensures TypeScript knows about vitest globals
