+++ apps/admin-x-framework/test/unit/filter-sources/create-remote-value-source.test.tsx
@@ -1,4 +1,4 @@
-import {createRemoteValueSource} from '@src/hooks/filter-sources/create-remote-value-source';
+import {createRemoteValueSource} from '../../../src/filter-sources/create-remote-value-source';
 import {describe, expect, it} from 'vitest';
 import {renderHook} from '@testing-library/react';
