+++ apps/admin-x-framework/test/unit/filters/filter-normalization.test.ts
@@ -1,5 +1,5 @@
 import {describe, expect, it} from 'vitest';
-import {escapeNqlString, formatDateInTimezone, getDayBoundsInUtc} from '@src/views/filters/filter-normalization';
+import {escapeNqlString, formatDateInTimezone, getDayBoundsInUtc} from '../../../src/filters/filter-normalization';
 
 describe('filter-normalization', () => {
     it('escapes single quotes for NQL strings', () => {
