+++ apps/admin-x-framework/test/unit/utils/get-site-timezone.test.ts
@@ -1,4 +1,4 @@
-import {getSiteTimezone} from '@src/utils/get-site-timezone';
+import {getSiteTimezone} from '../../../src/utils/get-site-timezone';
 
 describe('getSiteTimezone', () => {
     it('returns the site timezone if found', () => {
