+++ apps/posts/src/views/comments/comments.tsx
@@ -7,8 +7,8 @@ import {FilterBar, PageHeader, createFilter} from '@tryghost/shade/patterns';
 import {ListPage} from '@tryghost/shade/page-templates';
 import {LucideIcon} from '@tryghost/shade/utils';
 import {adminCommentIncludes, useBrowseComments} from '@tryghost/admin-x-framework/api/comments';
-import {escapeNqlString} from '../filters/filter-normalization';
-import {getSiteTimezone} from '@src/utils/get-site-timezone';
+import {escapeNqlString} from '@tryghost/admin-x-framework/filters';
+import {getSiteTimezone} from '@tryghost/admin-x-framework/utils/get-site-timezone';
 import {serializeCommentFilters} from './comment-filter-query';
 import {shouldDelayCommentDateFilterHydration, useFilterState} from './hooks/use-filter-state';
 import {useBrowseSettings} from '@tryghost/admin-x-framework/api/settings';
