+++ apps/posts/src/views/members/member-fields.ts
@@ -1,11 +1,5 @@
-import {DATE_FILTER_OPERATORS, DEFAULT_DATE_OPERATOR} from '../filters/filter-date';
+import {DATE_FILTER_OPERATORS, DEFAULT_DATE_OPERATOR, type FilterCodec, dateCodec, defineFields, escapeNqlString, extractComparator, numberCodec, scalarCodec, setCodec, textCodec, withFutureRelativeOperator, withPastRelativeOperator} from '@tryghost/admin-x-framework/filters';
 import {MULTIPLE_ACTIVE_STRIPE_CUSTOMERS_FIELD, MULTIPLE_ACTIVE_STRIPE_CUSTOMERS_FILTER, NO_MULTIPLE_ACTIVE_STRIPE_CUSTOMERS_FILTER} from './multiple-active-subscriptions';
-import {dateCodec, numberCodec, scalarCodec, setCodec, textCodec} from '../filters/filter-codecs';
-import {defineFields} from '../filters/filter-types';
-import {escapeNqlString} from '../filters/filter-normalization';
-import {extractComparator} from '../filters/filter-ast';
-import {withFutureRelativeOperator, withPastRelativeOperator} from '../filters/filter-relative-date';
-import type {FilterCodec} from '../filters/filter-types';
 
 const TEXT_OPERATORS = ['is', 'contains', 'does-not-contain', 'starts-with', 'ends-with'] as const;
 const NUMBER_OPERATORS = ['is', 'is-greater', 'is-less'] as const;
