+++ apps/posts/src/views/members/use-member-filter-fields.ts
@@ -1,13 +1,10 @@
 import React, {useMemo} from 'react';
-import {DATE_OPERATOR_LABELS} from '../filters/filter-date';
+import {DATE_OPERATOR_LABELS, RELATIVE_DATE_OPERATOR_LABELS, createOperatorOptions, createRelativeDateRenderer, fieldHasRelativeOperator, getTodayInTimezone} from '@tryghost/admin-x-framework/filters';
 import {FilterFieldConfig, FilterFieldGroup, FilterOption, ValueSource} from '@tryghost/shade/patterns';
 import {LabelFilterRenderer} from '@src/components/label-picker';
 import {LucideIcon} from '@tryghost/shade/utils';
 import {MULTIPLE_ACTIVE_STRIPE_CUSTOMERS_FIELD} from './multiple-active-subscriptions';
-import {RELATIVE_DATE_OPERATOR_LABELS, createRelativeDateRenderer, fieldHasRelativeOperator} from '../filters/filter-relative-date';
-import {createOperatorOptions} from '../filters/filter-operator-options';
 import {getMemberFields} from './member-fields';
-import {getTodayInTimezone} from '../filters/filter-normalization';
 import type {Offer} from '@tryghost/admin-x-framework/api/offers';
 
 interface UseMemberFilterFieldsOptions {
